Compare underlying enum types using `getDisplayString`
I'm not sure if this is correct, so I added some TODOs.
This is a problem if somehow the display strings are the same, e.g.
if two symbols of the same name come from different libraries.

This fixes the case where an enum is nullable but has a defaultsTo
value, where the default value is the non-nullable enum but the
field is nullable.